+++ docs/misc/code/cdq-divide/cdq-divide_3.cpp
@@ -1,241 +1,164 @@
+// 一道二维最长上升子序列的题
+// 为了确定某一个元素是否在最长上升子序列中可以正反跑两遍 CDQ
 #include <algorithm>
 #include <cstdio>
-#include <map>
-#include <set>
-#define SNI set<nod>::iterator
-#define SDI set<data>::iterator
 using namespace std;
-const int N = 1e5 + 10;
-int n;
-int m;
-int pre[N];
-int npre[N];
-int a[N];
-int tp[N];
-int lf[N];
-int rt[N];
-int co[N];
+typedef double db;
+const int N = 1e6 + 10;
 
-struct modi {
-  int t;
-  int pos;
-  int pre;
-  int va;
-
-  friend bool operator<(modi a, modi b) { return a.pre < b.pre; }
-} md[10 * N];
-
-int tp1;
-
-struct qry {
-  int t;
-  int l;
-  int r;
-  int ans;
-
-  friend bool operator<(qry a, qry b) { return a.l < b.l; }
-} qr[N];
-
-int tp2;
-int cnt;
-
-inline bool cmp(const qry& a, const qry& b) { return a.t < b.t; }
-
-inline void modify(int pos, int co)  // 修改函数
-{
-  if (npre[pos] == co) return;
-  md[++tp1] = (modi){++cnt, pos, npre[pos], -1};
-  md[++tp1] = (modi){++cnt, pos, npre[pos] = co, 1};
-}
-
-namespace prew {
-int lst[2 * N];
-map<int, int> mp;  // 提前离散化
-
-inline void prew() {
-  scanf("%d%d", &n, &m);
-  for (int i = 1; i <= n; i++) scanf("%d", &a[i]), mp[a[i]] = 1;
-  for (int i = 1; i <= m; i++) {
-    scanf("%d%d%d", &tp[i], &lf[i], &rt[i]);
-    if (tp[i] == 1) scanf("%d", &co[i]), mp[co[i]] = 1;
-  }
-  map<int, int>::iterator it, it1;
-  for (it = mp.begin(), it1 = it, ++it1; it1 != mp.end(); ++it, ++it1)
-    it1->second += it->second;
-  for (int i = 1; i <= n; i++) a[i] = mp[a[i]];
-  for (int i = 1; i <= n; i++)
-    if (tp[i] == 1) co[i] = mp[co[i]];
-  for (int i = 1; i <= n; i++) pre[i] = lst[a[i]], lst[a[i]] = i;
-  for (int i = 1; i <= n; i++) npre[i] = pre[i];
-}
-}  // namespace prew
-
-namespace colist {
 struct data {
-  int l;
-  int r;
-  int x;
+  int h;
+  int v;
+  int p;
+  int ma;
+  db ca;
+} a[2][N];
 
-  friend bool operator<(data a, data b) { return a.r < b.r; }
-};
+int n;
+bool tr;
 
-set<data> s;
-
-struct nod {
-  int l;
-  int r;
-
-  friend bool operator<(nod a, nod b) { return a.r < b.r; }
-};
-
-set<nod> c[2 * N];
-set<int> bd;
-
-inline void split(int mid) {  // 将一个节点拆成两个节点
-  SDI it = s.lower_bound((data){0, mid, 0});
-  data p = *it;
-  if (mid == p.r) return;
-  s.erase(p);
-  s.insert((data){p.l, mid, p.x});
-  s.insert((data){mid + 1, p.r, p.x});
-  c[p.x].erase((nod){p.l, p.r});
-  c[p.x].insert((nod){p.l, mid});
-  c[p.x].insert((nod){mid + 1, p.r});
+// 底下是重写比较
+inline bool cmp1(const data& a, const data& b) {
+  if (tr)
+    return a.h > b.h;
+  else
+    return a.h < b.h;
 }
 
-inline void del(set<data>::iterator it) {  // 删除一个迭代器
-  bd.insert(it->l);
-  SNI it1, it2;
-  it1 = it2 = c[it->x].find((nod){it->l, it->r});
-  ++it2;
-  if (it2 != c[it->x].end()) bd.insert(it2->l);
-  c[it->x].erase(it1);
-  s.erase(it);
+inline bool cmp2(const data& a, const data& b) {
+  if (tr)
+    return a.v > b.v;
+  else
+    return a.v < b.v;
 }
 
-inline void ins(data p) {  // 插入一个节点
-  s.insert(p);
-  SNI it = c[p.x].insert((nod){p.l, p.r}).first;
-  ++it;
-  if (it != c[p.x].end()) {
-    bd.insert(it->l);
-  }
+inline bool cmp3(const data& a, const data& b) {
+  if (tr)
+    return a.p < b.p;
+  else
+    return a.p > b.p;
 }
 
-inline void stv(int l, int r, int x) {  // 区间赋值
-  if (l != 1) split(l - 1);
-  split(r);
-  int p = l;  // split两下之后删掉所有区间
-  while (p != r + 1) {
-    SDI it = s.lower_bound((data){0, p, 0});
-    p = it->r + 1;
-    del(it);
-  }
-  ins((data){l, r, x});  // 扫一遍set处理所有变化的pre值
-  for (set<int>::iterator it = bd.begin(); it != bd.end(); ++it) {
-    SDI it1 = s.lower_bound((data){0, *it, 0});
-    if (*it != it1->l)
-      modify(*it, *it - 1);
-    else {
-      SNI it2 = c[it1->x].lower_bound((nod){0, *it});
-      if (it2 != c[it1->x].begin())
-        --it2, modify(*it, it2->r);
-      else
-        modify(*it, 0);
+inline bool cmp4(const data& a, const data& b) { return a.v == b.v; }
+
+struct treearray {
+  int ma[2 * N];
+  db ca[2 * N];
+
+  inline void c(int x, int t, db c) {
+    for (; x <= n; x += x & (-x)) {
+      if (ma[x] == t) {
+        ca[x] += c;
+      } else if (ma[x] < t) {
+        ca[x] = c;
+        ma[x] = t;
+      }
     }
-  }
-  bd.clear();
-}
-
-inline void ih() {
-  int nc = a[1];
-  int ccnt = 1;  // 将连续的一段插入到set中
-  for (int i = 2; i <= n; i++)
-    if (nc != a[i]) {
-      s.insert((data){i - ccnt, i - 1, nc}),
-          c[nc].insert((nod){i - ccnt, i - 1});
-      nc = a[i];
-      ccnt = 1;
-    } else {
-      ccnt++;
-    }
-  s.insert((data){n - ccnt + 1, n, a[n]}),
-      c[a[n]].insert((nod){n - ccnt + 1, n});
-}
-}  // namespace colist
-
-namespace CDQ {
-struct treearray  // 树状数组
-{
-  int ta[N];
-
-  inline void c(int x, int t) {
-    for (; x <= n; x += x & (-x)) ta[x] += t;
   }
 
   inline void d(int x) {
-    for (; x <= n; x += x & (-x)) ta[x] = 0;
+    for (; x <= n; x += x & (-x)) {
+      ma[x] = 0;
+      ca[x] = 0;
+    }
   }
 
-  inline int q(int x) {
-    int r = 0;
-    for (; x; x -= x & (-x)) r += ta[x];
-    return r;
-  }
-
-  inline void clear() {
-    for (int i = 1; i <= n; i++) ta[i] = 0;
+  inline void q(int x, int& m, db& c) {
+    for (; x > 0; x -= x & (-x)) {
+      if (ma[x] == m) {
+        c += ca[x];
+      } else if (m < ma[x]) {
+        c = ca[x];
+        m = ma[x];
+      }
+    }
   }
 } ta;
 
-int srt[N];
+int rk[2][N];
 
-inline bool cmp1(const int& a, const int& b) { return pre[a] < pre[b]; }
-
-inline void solve(int l1, int r1, int l2, int r2, int L, int R) {  // CDQ
-  if (l1 == r1 || l2 == r2) return;
-  int mid = (L + R) / 2;
-  int mid1 = l1;
-  while (mid1 != r1 && md[mid1 + 1].t <= mid) mid1++;
-  int mid2 = l2;
-  while (mid2 != r2 && qr[mid2 + 1].t <= mid) mid2++;
-  solve(l1, mid1, l2, mid2, L, mid);
-  solve(mid1, r1, mid2, r2, mid, R);
-  if (l1 != mid1 && mid2 != r2) {
-    sort(md + l1 + 1, md + mid1 + 1);
-    sort(qr + mid2 + 1, qr + r2 + 1);
-    for (int i = mid2 + 1, j = l1 + 1; i <= r2; i++) {  // 考虑左侧对右侧贡献
-      while (j <= mid1 && md[j].pre < qr[i].l) ta.c(md[j].pos, md[j].va), j++;
-      qr[i].ans += ta.q(qr[i].r) - ta.q(qr[i].l - 1);
+inline void solve(int l, int r, int t) {  // 递归跑
+  if (r - l == 1) {
+    return;
+  }
+  int mid = (l + r) / 2;
+  solve(l, mid, t);
+  sort(a[t] + mid + 1, a[t] + r + 1, cmp1);
+  int p = l + 1;
+  for (int i = mid + 1; i <= r; i++) {
+    for (; (cmp1(a[t][p], a[t][i]) || a[t][p].h == a[t][i].h) && p <= mid;
+         p++) {
+      ta.c(a[t][p].v, a[t][p].ma, a[t][p].ca);
     }
-    for (int i = l1 + 1; i <= mid1; i++) ta.d(md[i].pos);
+    db c = 0;
+    int m = 0;
+    ta.q(a[t][i].v, m, c);
+    if (a[t][i].ma < m + 1) {
+      a[t][i].ma = m + 1;
+      a[t][i].ca = c;
+    } else if (a[t][i].ma == m + 1) {
+      a[t][i].ca += c;
+    }
+  }
+  for (int i = l + 1; i <= mid; i++) {
+    ta.d(a[t][i].v);
+  }
+  sort(a[t] + mid, a[t] + r + 1, cmp3);
+  solve(mid, r, t);
+  sort(a[t] + l + 1, a[t] + r + 1, cmp1);
+}
+
+inline void ih(int t) {
+  sort(a[t] + 1, a[t] + n + 1, cmp2);
+  rk[t][1] = 1;
+  for (int i = 2; i <= n; i++) {
+    rk[t][i] = (cmp4(a[t][i], a[t][i - 1])) ? rk[t][i - 1] : i;
+  }
+  for (int i = 1; i <= n; i++) {
+    a[t][i].v = rk[t][i];
+  }
+  sort(a[t] + 1, a[t] + n + 1, cmp3);
+  for (int i = 1; i <= n; i++) {
+    a[t][i].ma = 1;
+    a[t][i].ca = 1;
   }
 }
 
-inline void mainsolve() {
-  colist::ih();
-  for (int i = 1; i <= m; i++)
-    if (tp[i] == 1)
-      colist::stv(lf[i], rt[i], co[i]);
-    else
-      qr[++tp2] = (qry){++cnt, lf[i], rt[i], 0};
-  sort(qr + 1, qr + tp2 + 1);
-  for (int i = 1; i <= n; i++) srt[i] = i;
-  sort(srt + 1, srt + n + 1, cmp1);
-  for (int i = 1, j = 1; i <= tp2; i++) {  // 初始化一下每个询问的值
-    while (j <= n && pre[srt[j]] < qr[i].l) ta.c(srt[j], 1), j++;
-    qr[i].ans += ta.q(qr[i].r) - ta.q(qr[i].l - 1);
-  }
-  ta.clear();
-  sort(qr + 1, qr + tp2 + 1, cmp);
-  solve(0, tp1, 0, tp2, 0, cnt);
-  sort(qr + 1, qr + tp2 + 1, cmp);
-  for (int i = 1; i <= tp2; i++) printf("%d\n", qr[i].ans);
-}
-}  // namespace CDQ
+int len;
+db ans;
 
 int main() {
-  prew::prew();
-  CDQ::mainsolve();
+  scanf("%d", &n);
+  for (int i = 1; i <= n; i++) {
+    scanf("%d%d", &a[0][i].h, &a[0][i].v);
+    a[0][i].p = i;
+    a[1][i].h = a[0][i].h;
+    a[1][i].v = a[0][i].v;
+    a[1][i].p = i;
+  }
+  ih(0);
+  solve(0, n, 0);
+  tr = 1;
+  ih(1);
+  solve(0, n, 1);
+  tr = 1;
+  sort(a[0] + 1, a[0] + n + 1, cmp3);
+  sort(a[1] + 1, a[1] + n + 1, cmp3);
+  for (int i = 1; i <= n; i++) {
+    len = max(len, a[0][i].ma);
+  }
+  printf("%d\n", len);
+  for (int i = 1; i <= n; i++) {
+    if (a[0][i].ma == len) {
+      ans += a[0][i].ca;
+    }
+  }
+  for (int i = 1; i <= n; i++) {
+    if (a[0][i].ma + a[1][i].ma - 1 == len) {
+      printf("%.5lf ", (a[0][i].ca * a[1][i].ca) / ans);
+    } else {
+      printf("0.00000 ");
+    }
+  }
   return 0;
 }
