+++ docs/misc/code/cdq-divide/cdq-divide_4.cpp
@@ -1,225 +1,241 @@
 #include <algorithm>
 #include <cstdio>
-#include <stack>
-#include <vector>
+#include <map>
+#include <set>
+#define SNI set<nod>::iterator
+#define SDI set<data>::iterator
 using namespace std;
-typedef long long ll;
+const int N = 1e5 + 10;
 int n;
 int m;
-int ask;
+int pre[N];
+int npre[N];
+int a[N];
+int tp[N];
+int lf[N];
+int rt[N];
+int co[N];
 
-struct bcj {
-  int fa[20010];
-  int size[20010];
+struct modi {
+  int t;
+  int pos;
+  int pre;
+  int va;
 
-  struct opt {
-    int u;
-    int v;
-  };
+  friend bool operator<(modi a, modi b) { return a.pre < b.pre; }
+} md[10 * N];
 
-  stack<opt> st;
+int tp1;
 
-  inline void ih() {
-    for (int i = 1; i <= n; i++) fa[i] = i, size[i] = 1;
-  }
+struct qry {
+  int t;
+  int l;
+  int r;
+  int ans;
 
-  inline int f(int x) { return (fa[x] == x) ? x : f(fa[x]); }
+  friend bool operator<(qry a, qry b) { return a.l < b.l; }
+} qr[N];
 
-  inline void u(int x, int y) {  // 带撤回
-    int u = f(x);
-    int v = f(y);
-    if (u == v) return;
-    if (size[u] < size[v]) swap(u, v);
-    size[u] += size[v];
-    fa[v] = u;
-    opt o;
-    o.u = u;
-    o.v = v;
-    st.push(o);
-  }
+int tp2;
+int cnt;
 
-  inline void undo() {
-    opt o = st.top();
-    st.pop();
-    fa[o.v] = o.v;
-    size[o.u] -= size[o.v];
-  }
+inline bool cmp(const qry& a, const qry& b) { return a.t < b.t; }
 
-  inline void clear(int tim) {
-    while (st.size() > tim) {
-      undo();
-    }
-  }
-} s, s1;
-
-struct edge  // 静态边
+inline void modify(int pos, int co)  // 修改函数
 {
-  int u;
-  int v;
-  ll val;
-  int mrk;
-
-  friend bool operator<(edge a, edge b) { return a.val < b.val; }
-} e[50010];
-
-struct moved {
-  int u;
-  int v;
-};  // 动态边
-
-struct query {
-  int num;
-  ll val;
-  ll ans;
-} q[50010];
-
-bool book[50010];  // 询问
-vector<edge> ve[30];
-vector<moved> vq;
-vector<edge> tr;
-ll res[30];
-int tim[30];
-
-inline void pushdown(int dep)  // 缩边
-{
-  tr.clear();  // 这里要复制一份，以免无法回撤操作
-  for (int i = 0; i < ve[dep].size(); i++) {
-    tr.push_back(ve[dep][i]);
-  }
-  sort(tr.begin(), tr.end());
-  for (int i = 0; i < tr.size(); i++) {  // 无用边
-    if (s1.f(tr[i].u) == s1.f(tr[i].v)) {
-      tr[i].mrk = -1;
-      continue;
-    }
-    s1.u(tr[i].u, tr[i].v);
-  }
-  s1.clear(0);
-  res[dep + 1] = res[dep];
-  for (int i = 0; i < vq.size(); i++) {
-    s1.u(vq[i].u, vq[i].v);
-  }
-  vq.clear();
-  for (int i = 0; i < tr.size(); i++) {  // 必须边
-    if (tr[i].mrk == -1 || s1.f(tr[i].u) == s1.f(tr[i].v)) continue;
-    tr[i].mrk = 1;
-    s1.u(tr[i].u, tr[i].v);
-    s.u(tr[i].u, tr[i].v);
-    res[dep + 1] += tr[i].val;
-  }
-  s1.clear(0);
-  ve[dep + 1].clear();
-  for (int i = 0; i < tr.size(); i++) {  // 缩边
-    if (tr[i].mrk != 0) continue;
-    edge p;
-    p.u = s.f(tr[i].u);
-    p.v = s.f(tr[i].v);
-    if (p.u == p.v) continue;
-    p.val = tr[i].val;
-    p.mrk = 0;
-    ve[dep + 1].push_back(p);
-  }
-  return;
+  if (npre[pos] == co) return;
+  md[++tp1] = (modi){++cnt, pos, npre[pos], -1};
+  md[++tp1] = (modi){++cnt, pos, npre[pos] = co, 1};
 }
 
-inline void solve(int l, int r, int dep) {
-  tim[dep] = s.st.size();
-  int mid = (l + r) / 2;
-  if (r - l == 1) {  // 终止条件
-    edge p;
-    p.u = s.f(e[q[r].num].u);
-    p.v = s.f(e[q[r].num].v);
-    p.val = q[r].val;
-    e[q[r].num].val = q[r].val;
-    p.mrk = 0;
-    ve[dep].push_back(p);
-    pushdown(dep);
-    q[r].ans = res[dep + 1];
-    s.clear(tim[dep - 1]);
-    return;
+namespace prew {
+int lst[2 * N];
+map<int, int> mp;  // 提前离散化
+
+inline void prew() {
+  scanf("%d%d", &n, &m);
+  for (int i = 1; i <= n; i++) scanf("%d", &a[i]), mp[a[i]] = 1;
+  for (int i = 1; i <= m; i++) {
+    scanf("%d%d%d", &tp[i], &lf[i], &rt[i]);
+    if (tp[i] == 1) scanf("%d", &co[i]), mp[co[i]] = 1;
   }
-  for (int i = l + 1; i <= mid; i++) {
-    book[q[i].num] = true;
-  }
-  for (int i = mid + 1; i <= r; i++) {  // 动转静
-    if (book[q[i].num]) continue;
-    edge p;
-    p.u = s.f(e[q[i].num].u);
-    p.v = s.f(e[q[i].num].v);
-    p.val = e[q[i].num].val;
-    p.mrk = 0;
-    ve[dep].push_back(p);
-  }
-  for (int i = l + 1; i <= mid; i++) {  // 询问转动态
-    moved p;
-    p.u = s.f(e[q[i].num].u);
-    p.v = s.f(e[q[i].num].v);
-    vq.push_back(p);
-  }
-  pushdown(dep);  // 下面的是回撤
-  for (int i = mid + 1; i <= r; i++) {
-    if (book[q[i].num]) continue;
-    ve[dep].pop_back();
-  }
-  for (int i = l + 1; i <= mid; i++) {
-    book[q[i].num] = false;
-  }
-  solve(l, mid, dep + 1);
-  for (int i = 0; i < ve[dep].size(); i++) {
-    ve[dep][i].mrk = 0;
-  }
-  for (int i = mid + 1; i <= r; i++) {
-    book[q[i].num] = true;
-  }
-  for (int i = l + 1; i <= mid; i++) {  // 动转静
-    if (book[q[i].num]) continue;
-    edge p;
-    p.u = s.f(e[q[i].num].u);
-    p.v = s.f(e[q[i].num].v);
-    p.val = e[q[i].num].val;
-    p.mrk = 0;
-    ve[dep].push_back(p);
-  }
-  for (int i = mid + 1; i <= r; i++) {  // 询问转动
-    book[q[i].num] = false;
-    moved p;
-    p.u = s.f(e[q[i].num].u);
-    p.v = s.f(e[q[i].num].v);
-    vq.push_back(p);
-  }
-  pushdown(dep);
-  solve(mid, r, dep + 1);
-  s.clear(tim[dep - 1]);
-  return;  // 时间倒流至上一层
+  map<int, int>::iterator it, it1;
+  for (it = mp.begin(), it1 = it, ++it1; it1 != mp.end(); ++it, ++it1)
+    it1->second += it->second;
+  for (int i = 1; i <= n; i++) a[i] = mp[a[i]];
+  for (int i = 1; i <= n; i++)
+    if (tp[i] == 1) co[i] = mp[co[i]];
+  for (int i = 1; i <= n; i++) pre[i] = lst[a[i]], lst[a[i]] = i;
+  for (int i = 1; i <= n; i++) npre[i] = pre[i];
 }
+}  // namespace prew
+
+namespace colist {
+struct data {
+  int l;
+  int r;
+  int x;
+
+  friend bool operator<(data a, data b) { return a.r < b.r; }
+};
+
+set<data> s;
+
+struct nod {
+  int l;
+  int r;
+
+  friend bool operator<(nod a, nod b) { return a.r < b.r; }
+};
+
+set<nod> c[2 * N];
+set<int> bd;
+
+inline void split(int mid) {  // 将一个节点拆成两个节点
+  SDI it = s.lower_bound((data){0, mid, 0});
+  data p = *it;
+  if (mid == p.r) return;
+  s.erase(p);
+  s.insert((data){p.l, mid, p.x});
+  s.insert((data){mid + 1, p.r, p.x});
+  c[p.x].erase((nod){p.l, p.r});
+  c[p.x].insert((nod){p.l, mid});
+  c[p.x].insert((nod){mid + 1, p.r});
+}
+
+inline void del(set<data>::iterator it) {  // 删除一个迭代器
+  bd.insert(it->l);
+  SNI it1, it2;
+  it1 = it2 = c[it->x].find((nod){it->l, it->r});
+  ++it2;
+  if (it2 != c[it->x].end()) bd.insert(it2->l);
+  c[it->x].erase(it1);
+  s.erase(it);
+}
+
+inline void ins(data p) {  // 插入一个节点
+  s.insert(p);
+  SNI it = c[p.x].insert((nod){p.l, p.r}).first;
+  ++it;
+  if (it != c[p.x].end()) {
+    bd.insert(it->l);
+  }
+}
+
+inline void stv(int l, int r, int x) {  // 区间赋值
+  if (l != 1) split(l - 1);
+  split(r);
+  int p = l;  // split两下之后删掉所有区间
+  while (p != r + 1) {
+    SDI it = s.lower_bound((data){0, p, 0});
+    p = it->r + 1;
+    del(it);
+  }
+  ins((data){l, r, x});  // 扫一遍set处理所有变化的pre值
+  for (set<int>::iterator it = bd.begin(); it != bd.end(); ++it) {
+    SDI it1 = s.lower_bound((data){0, *it, 0});
+    if (*it != it1->l)
+      modify(*it, *it - 1);
+    else {
+      SNI it2 = c[it1->x].lower_bound((nod){0, *it});
+      if (it2 != c[it1->x].begin())
+        --it2, modify(*it, it2->r);
+      else
+        modify(*it, 0);
+    }
+  }
+  bd.clear();
+}
+
+inline void ih() {
+  int nc = a[1];
+  int ccnt = 1;  // 将连续的一段插入到set中
+  for (int i = 2; i <= n; i++)
+    if (nc != a[i]) {
+      s.insert((data){i - ccnt, i - 1, nc}),
+          c[nc].insert((nod){i - ccnt, i - 1});
+      nc = a[i];
+      ccnt = 1;
+    } else {
+      ccnt++;
+    }
+  s.insert((data){n - ccnt + 1, n, a[n]}),
+      c[a[n]].insert((nod){n - ccnt + 1, n});
+}
+}  // namespace colist
+
+namespace CDQ {
+struct treearray  // 树状数组
+{
+  int ta[N];
+
+  inline void c(int x, int t) {
+    for (; x <= n; x += x & (-x)) ta[x] += t;
+  }
+
+  inline void d(int x) {
+    for (; x <= n; x += x & (-x)) ta[x] = 0;
+  }
+
+  inline int q(int x) {
+    int r = 0;
+    for (; x; x -= x & (-x)) r += ta[x];
+    return r;
+  }
+
+  inline void clear() {
+    for (int i = 1; i <= n; i++) ta[i] = 0;
+  }
+} ta;
+
+int srt[N];
+
+inline bool cmp1(const int& a, const int& b) { return pre[a] < pre[b]; }
+
+inline void solve(int l1, int r1, int l2, int r2, int L, int R) {  // CDQ
+  if (l1 == r1 || l2 == r2) return;
+  int mid = (L + R) / 2;
+  int mid1 = l1;
+  while (mid1 != r1 && md[mid1 + 1].t <= mid) mid1++;
+  int mid2 = l2;
+  while (mid2 != r2 && qr[mid2 + 1].t <= mid) mid2++;
+  solve(l1, mid1, l2, mid2, L, mid);
+  solve(mid1, r1, mid2, r2, mid, R);
+  if (l1 != mid1 && mid2 != r2) {
+    sort(md + l1 + 1, md + mid1 + 1);
+    sort(qr + mid2 + 1, qr + r2 + 1);
+    for (int i = mid2 + 1, j = l1 + 1; i <= r2; i++) {  // 考虑左侧对右侧贡献
+      while (j <= mid1 && md[j].pre < qr[i].l) ta.c(md[j].pos, md[j].va), j++;
+      qr[i].ans += ta.q(qr[i].r) - ta.q(qr[i].l - 1);
+    }
+    for (int i = l1 + 1; i <= mid1; i++) ta.d(md[i].pos);
+  }
+}
+
+inline void mainsolve() {
+  colist::ih();
+  for (int i = 1; i <= m; i++)
+    if (tp[i] == 1)
+      colist::stv(lf[i], rt[i], co[i]);
+    else
+      qr[++tp2] = (qry){++cnt, lf[i], rt[i], 0};
+  sort(qr + 1, qr + tp2 + 1);
+  for (int i = 1; i <= n; i++) srt[i] = i;
+  sort(srt + 1, srt + n + 1, cmp1);
+  for (int i = 1, j = 1; i <= tp2; i++) {  // 初始化一下每个询问的值
+    while (j <= n && pre[srt[j]] < qr[i].l) ta.c(srt[j], 1), j++;
+    qr[i].ans += ta.q(qr[i].r) - ta.q(qr[i].l - 1);
+  }
+  ta.clear();
+  sort(qr + 1, qr + tp2 + 1, cmp);
+  solve(0, tp1, 0, tp2, 0, cnt);
+  sort(qr + 1, qr + tp2 + 1, cmp);
+  for (int i = 1; i <= tp2; i++) printf("%d\n", qr[i].ans);
+}
+}  // namespace CDQ
 
 int main() {
-  scanf("%d%d%d", &n, &m, &ask);
-  s.ih();
-  s1.ih();
-  for (int i = 1; i <= m; i++) {
-    scanf("%d%d%lld", &e[i].u, &e[i].v, &e[i].val);
-  }
-  for (int i = 1; i <= ask; i++) {
-    scanf("%d%lld", &q[i].num, &q[i].val);
-  }
-  for (int i = 1; i <= ask; i++) {  // 初始动态边
-    book[q[i].num] = true;
-    moved p;
-    p.u = e[q[i].num].u;
-    p.v = e[q[i].num].v;
-    vq.push_back(p);
-  }
-  for (int i = 1; i <= m; i++) {
-    if (book[i]) continue;
-    ve[1].push_back(e[i]);
-  }  // 初始静态
-  for (int i = 1; i <= ask; i++) {
-    book[q[i].num] = false;
-  }
-  solve(0, ask, 1);
-  for (int i = 1; i <= ask; i++) {
-    printf("%lld\n", q[i].ans);
-  }
+  prew::prew();
+  CDQ::mainsolve();
   return 0;
 }
