+++ docs/misc/examples/cdq-divide/cdq-divide_1.ans
@@ -1,4 +1,10 @@
-5
-2
-2
+3
+1
+3
+0
+1
+0
+1
+0
+0
 1
