+++ docs/misc/examples/cdq-divide/cdq-divide_1.in
@@ -1,10 +1,11 @@
-5 4
-1
-5
-3
-4
-2
-5
-1
-4
-2
+10 3
+3 3 3
+2 3 3
+2 3 1
+3 1 1
+3 1 2
+1 3 1
+1 1 2
+1 2 2
+1 3 2
+1 2 1
