+++ docs/misc/examples/cdq-divide/cdq-divide_2.ans
@@ -1,2 +1,4 @@
+5
 2
-0.33333 0.33333 0.33333 1.00000
+2
+1
