+++ docs/misc/examples/cdq-divide/cdq-divide_2.in
@@ -1,5 +1,10 @@
+5 4
+1
+5
+3
 4
-3 30
-4 40
-6 60
-3 30
+2
+5
+1
+4
+2
