+++ docs/misc/examples/cdq-divide/cdq-divide_3.ans
@@ -1,4 +1,2 @@
-5
-3
-1
-1
+2
+0.33333 0.33333 0.33333 1.00000
