+++ docs/misc/examples/cdq-divide/cdq-divide_3.in
@@ -1,7 +1,5 @@
-5 5
-1 2 3 4 5
-2 1 5
-1 2 3 4
-2 1 5
-2 3 3
-2 2 4
+4
+3 30
+4 40
+6 60
+3 30
