+++ docs/misc/examples/cdq-divide/cdq-divide_4.ans
@@ -1,3 +1,4 @@
-14
-10
-9
+5
+3
+1
+1
