+++ docs/misc/examples/cdq-divide/cdq-divide_4.in
@@ -1,9 +1,7 @@
-5 5 3
-1 2 1
-2 3 2
-3 4 3
-4 5 4
-5 1 5
-1 6
-1 1
-5 3
+5 5
+1 2 3 4 5
+2 1 5
+1 2 3 4
+2 1 5
+2 3 3
+2 2 4
