+++ docs/misc/examples/cdq-divide/cdq-divide_5.in
@@ -1,0 +1,9 @@
+5 5 3
+1 2 1
+2 3 2
+3 4 3
+4 5 4
+5 1 5
+1 6
+1 1
+5 3
